[REM] Product: Odoo doesn't pass the default_code to a product_product
When this feature was created, the intention was to prevent a product from being created without a default_code, however, it was detected that when creating a product.product, this field is not shared when it is made by a product.template, which now creates a conflict when the user are trying to create a new product. To fix it, an automated action has been created that will replace this feature.
But also, was modified the create and copy functions to add the default_code of the product.template in product.product, and change the default code for a generic default_code in a copy.